backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.5
django-stubs==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt
Faker==36.1.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.15
langchain-core==0.3.31
langchain-google-genai
langchain-ollama==0.2.2
langchain-openai
langchain-text-splitters==0.3.5
langsmith==0.2.11
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.23.0
numpy==2.2.2
ollama==0.4.6
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pika==1.3.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pypdfium2==4.30.1
PySocks==1.7.1
pytest==8.3.4
pytest-django==4.10.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
selenium==4.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.5
SQLAlchemy==2.0.37
sqlparse==0.5.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
dj_database_url
gunicorn
google-cloud-secret-manager==2.21.1
django-environ==0.11.2
